Integration develop nr 2019w42

See merge request oai/openairinterface5g!691

+++ ci-scripts/buildOnVM.sh
@@ -44,7 +44,7 @@ function build_on_vm {
         STATUS=1
         return
     fi
-    if [ ! -f /etc/apt/apt.conf.d/01proxy ]
+    if [[ ! -f /etc/apt/apt.conf.d/01proxy ]] && [[ "$OPTIONAL_APTCACHER" != "true" ]]
     then
         echo "Missing /etc/apt/apt.conf.d/01proxy file!"
         echo "Is apt-cacher installed and configured?"
@@ -96,12 +96,12 @@ function build_on_vm {
     else
         scp -o StrictHostKeyChecking=no $JENKINS_WKSP/localZip.zip ubuntu@$VM_IP_ADDR:/home/ubuntu
     fi
-    scp -o StrictHostKeyChecking=no /etc/apt/apt.conf.d/01proxy ubuntu@$VM_IP_ADDR:/home/ubuntu
+    [ -f /etc/apt/apt.conf.d/01proxy ] && scp -o StrictHostKeyChecking=no /etc/apt/apt.conf.d/01proxy ubuntu@$VM_IP_ADDR:/home/ubuntu
 
     echo "############################################################"
     echo "Running install and build script on VM ($VM_NAME)"
     echo "############################################################"
-    echo "sudo cp 01proxy /etc/apt/apt.conf.d/" > $VM_CMDS
+    echo "[ -f 01proxy ] && sudo cp 01proxy /etc/apt/apt.conf.d/" > $VM_CMDS
     echo "touch /home/ubuntu/.hushlogin" >> $VM_CMDS
     echo "git config --global https.postBuffer 123289600" >> $VM_CMDS
     if [[ "$VM_NAME" == *"-cppcheck"* ]]
